Centralize IDP construction and make IDP comparison smarter
