+++ cmd/ceremony/crl.go
@@ -3,14 +3,13 @@ package main
 import (
 	"crypto"
 	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/asn1"
 	"encoding/pem"
 	"errors"
 	"fmt"
 	"math/big"
 	"time"
 
+	"github.com/letsencrypt/boulder/crl/idp"
 	"github.com/letsencrypt/boulder/linter"
 )
 
@@ -37,7 +36,7 @@ func generateCRL(signer crypto.Signer, issuer *x509.Certificate, thisUpdate, nex
 		return nil, errors.New("nextUpdate must be less than 12 months after thisUpdate")
 	}
 	// Add the Issuing Distribution Point extension.
-	idp, err := makeIDPExt()
+	idp, err := idp.MakeCACertsExt()
 	if err != nil {
 		return nil, fmt.Errorf("creating IDP extension: %w", err)
 	}
@@ -59,31 +58,4 @@ func generateCRL(signer crypto.Signer, issuer *x509.Certificate, thisUpdate, nex
 	}
 
 	return pem.EncodeToMemory(&pem.Block{Type: "X509 CRL", Bytes: crlBytes}), nil
-}
-
-// issuingDistributionPoint represents the ASN.1 IssuingDistributionPoint
-// SEQUENCE as defined in RFC 5280 Section 5.2.5. We only use one of the fields,
-// all others are omitted.
-// https://datatracker.ietf.org/doc/html/rfc5280#page-66
-type issuingDistributionPoint struct {
-	OnlyContainsCACerts bool `asn1:"optional,tag:2"`
-}
-
-// makeIDPExt returns a critical IssuingDistributionPoint extension enabling the
-// OnlyContainsCACerts boolean.
-func makeIDPExt() (*pkix.Extension, error) {
-	val := issuingDistributionPoint{
-		OnlyContainsCACerts: true,
-	}
-
-	valBytes, err := asn1.Marshal(val)
-	if err != nil {
-		return nil, err
-	}
-
-	return &pkix.Extension{
-		Id:       asn1.ObjectIdentifier{2, 5, 29, 28}, // id-ce-issuingDistributionPoint
-		Value:    valBytes,
-		Critical: true,
-	}, nil
 }
