+++ issuance/crl.go
@@ -3,8 +3,6 @@ package issuance
 import (
 	"crypto/rand"
 	"crypto/x509"
-	"crypto/x509/pkix"
-	"encoding/asn1"
 	"fmt"
 	"math/big"
 	"time"
@@ -12,6 +10,7 @@ import (
 	"github.com/zmap/zlint/v3/lint"
 
 	"github.com/letsencrypt/boulder/config"
+	"github.com/letsencrypt/boulder/crl/idp"
 	"github.com/letsencrypt/boulder/linter"
 )
 
@@ -92,7 +91,7 @@ func (i *Issuer) IssueCRL(prof *CRLProfile, req *CRLRequest) ([]byte, error) {
 		// contain the new-style CRLDP URL instead.
 		idps = append(idps, fmt.Sprintf("%s/%d/%d.crl", req.DeprecatedIDPBaseURL, i.NameID(), req.Shard))
 	}
-	idp, err := makeIDPExt(idps)
+	idp, err := idp.MakeUserCertsExt(idps)
 	if err != nil {
 		return nil, fmt.Errorf("creating IDP extension: %w", err)
 	}
@@ -114,52 +113,4 @@ func (i *Issuer) IssueCRL(prof *CRLProfile, req *CRLRequest) ([]byte, error) {
 	}
 
 	return crlBytes, nil
-}
-
-// distributionPointName represents the ASN.1 DistributionPointName CHOICE as
-// defined in RFC 5280 Section 4.2.1.13. We only use one of the fields, so the
-// others are omitted.
-type distributionPointName struct {
-	// Technically, FullName is of type GeneralNames, which is of type SEQUENCE OF
-	// GeneralName. But GeneralName itself is of type CHOICE, and the ans1.Marhsal
-	// function doesn't support marshalling structs to CHOICEs, so we have to use
-	// asn1.RawValue and encode the GeneralName ourselves.
-	FullName []asn1.RawValue `asn1:"optional,tag:0"`
-}
-
-// issuingDistributionPoint represents the ASN.1 IssuingDistributionPoint
-// SEQUENCE as defined in RFC 5280 Section 5.2.5. We only use two of the fields,
-// so the others are omitted.
-type issuingDistributionPoint struct {
-	DistributionPoint     distributionPointName `asn1:"optional,tag:0"`
-	OnlyContainsUserCerts bool                  `asn1:"optional,tag:1"`
-}
-
-// makeIDPExt returns a critical IssuingDistributionPoint extension containing
-// the given URLs and with the OnlyContainsUserCerts boolean set to true.
-func makeIDPExt(urls []string) (pkix.Extension, error) {
-	var gns []asn1.RawValue
-	for _, url := range urls {
-		gns = append(gns, asn1.RawValue{ // GeneralName
-			Class: 2, // context-specific
-			Tag:   6, // uniformResourceIdentifier, IA5String
-			Bytes: []byte(url),
-		})
-	}
-
-	val := issuingDistributionPoint{
-		DistributionPoint:     distributionPointName{FullName: gns},
-		OnlyContainsUserCerts: true,
-	}
-
-	valBytes, err := asn1.Marshal(val)
-	if err != nil {
-		return pkix.Extension{}, err
-	}
-
-	return pkix.Extension{
-		Id:       asn1.ObjectIdentifier{2, 5, 29, 28}, // id-ce-issuingDistributionPoint
-		Value:    valBytes,
-		Critical: true,
-	}, nil
 }
